added script docs
