+++ scripts/1-get-stations.js
