+++ scripts/3-find-near-stations.js
